Move shared types to fix type imports
