+++ packages/transform-dataresource/src/charts/shared.tsx
@@ -1,4 +1,4 @@
-import * as Dx from "Dx";
+import * as Dx from "../types";
 
 function stringOrFnAccessor(
   datapoint: Dx.Datapoint,
